Time: 15 ms (22.70%) | Memory: 23 MB (41.93%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -13,7 +13,7 @@ public:
         long long totalH = 0;
         int n = piles.size();
         for (int i = 0; i<n; i++){
-            totalH += (long long) ceil((double)piles[i]/(double)mid);
+            totalH += ceil((double)piles[i]/(double)mid);
         }
         return totalH;
     }
